# mapaviajero-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel App API",
        "description": "Documentación de la API Travel App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/tipoLugares": {
            "get": {
                "tags": [
                    "TipoLugares"
                ],
                "summary": "Listar todos los tipos de lugares",
                "description": "Obtiene la lista de todos los tipos de lugares disponibles. Solo accesible para usuarios registrados",
                "operationId": "d2df426e63216ec00b0768e98e834153",
                "responses": {
                    "200": {
                        "description": "Lista de tipos de lugares obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TipoLugar"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TipoLugares"
                ],
                "summary": "Crear un nuevo tipo de lugar",
                "description": "Solo accesible para administradores y superadministradores. Permite crear y añadir un nuevo tipo de lugar a la Base de Datos.",
                "operationId": "49d581769f1878d8ca6e7e687a67d448",
                "requestBody": {
                    "description": "Datos para crear un tipo de lugar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoLugarDB"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de lugar creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TipoLugar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo los administradores pueden guardar un nuevo tipo de lugar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The nombre has already been taken."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tipoLugares/{id}": {
            "get": {
                "tags": [
                    "TipoLugares"
                ],
                "summary": "Mostrar un tipo de lugar específico (por id)",
                "description": "Obtiene los detalles de un tipo de lugar por su ID. Solo accesible para usuarios registrados",
                "operationId": "9b32c2fec6d9b829527a47a6b40c1d2b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de lugar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de lugar encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TipoLugar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de lugar no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de lugar no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "TipoLugares"
                ],
                "summary": "Actualizar un tipo de lugar",
                "description": "Permite modificar los datos de un tipo de lugar. Solo accesible para administradores y superadministradores.",
                "operationId": "f6e8675e07a517ec15a9d21b871b0e5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de lugar a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoLugarDB"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de lugar actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TipoLugar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo el administrador puede editar un tipo e lugar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de lugar no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de lugar no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ya existe un tipo de lugar con ese nombre en nuestra BBDD."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "TipoLugares"
                ],
                "summary": "Eliminar un tipo de lugar",
                "description": "Permite eliminar un tipo de lugar. Solo accesible para administradores y superadministradores.",
                "operationId": "ff335a2f72e3daee3ea6724f83cb5e0e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de lugar a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de lugar borrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de lugar borrado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo el administrador puede borrar un tipo de lugar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de lugar no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de lugar no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TipoLugar": {
                "title": "Tipo de Lugar (Recurso API)",
                "description": "Recurso que representa un tipo de lugar (por ejemplo, playa, montaña, museo, etc.)",
                "properties": {
                    "id": {
                        "description": "ID único del tipo de lugar",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre del tipo de lugar",
                        "type": "string",
                        "example": "Playa"
                    }
                },
                "type": "object"
            },
...
            "TipoLugarDB": {
                "title": "Tipo de Lugar (BBDD)",
                "description": "Esquema del tipo de lugar tal y como está almacenado en la base de datos",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Alta montaña"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}